requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0

# ML
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.0

# Deep Learning
torch==2.3.0
torchmetrics==1.4.0

# MLOps
mlflow==2.13.0
evidently==0.4.30

# Serving
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.1

# Visualization
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# Utilities
PyYAML==6.0.1
tqdm==4.66.4
pyarrow==16.1.0
openpyxl==3.1.3

# Testing
pytest==8.2.1
pytest-cov==5.0.0
